requirements.txt
# Blood Smear Domain Expert - Requirements
# Python 3.8+ required

# Core deep learning frameworks
torch>=2.0.0
torchvision>=0.15.0

# Computer vision models
ultralytics>=8.0.0
timm>=0.9.0

# Image processing
Pillow>=9.0.0
opencv-python>=4.7.0

# Scientific computing
numpy>=1.24.0
scipy>=1.10.0

# Machine learning utilities
scikit-learn>=1.3.0

# NLP and embeddings
sentence-transformers>=2.2.0

# Vector database (with fallback to numpy)
chromadb>=0.4.0

# PDF processing
pypdf>=3.0.0

# LLM API
openai>=1.0.0

# Configuration
pyyaml>=6.0.0
python-dotenv>=1.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress bars
tqdm>=4.65.0

# HTTP requests (for PDF downloading)
requests>=2.28.0

# Development and testing (optional)
pytest>=7.3.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0

# Jupyter support (optional - for notebook compatibility)
jupyter>=1.0.0
ipywidgets>=8.0.0
